Refactor: swagger 설정 구조 정리
